Refactor for cleaner arch and removed unnecessary mixin classes
